Refactor files & functions NB: Breaking change to the module's public API - getTime (returning a timestamp) is replaced by getDatetime (return a Date object)
